requirements.txt
amqp==2.3.2
asn1crypto==0.24.0
bcrypt==3.1.4
billiard==3.5.0.4
captcha==0.3
celery==4.2.1
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
Click==7.0
command-not-found==0.3
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
cryptography==2.1.4
diff-match-patch==20121119
distro-info==0.18
Django==2.0.5
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-filter==1.1.0
django-filters==0.2.1
django-formtools==2.1
django-guardian==1.4.9
django-import-export==1.1.0
django-niji==0.1.0
django-pure-pagination==0.3.0
django-ranged-response==0.2.0
django-reversion==3.0.1
django-simple-captcha==0.5.9
djangorestframework==3.8.2
djangorestframework-filters==0.10.2.post0
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
future==0.17.0
httplib2==0.9.2
idna==2.6
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
keyring==10.6.0
keyrings.alt==3.0
kombu==4.2.1
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.0.1
MarkupSafe==1.0
mistune==0.8.3
oauth==1.0.1
odfpy==1.3.6
olefile==0.45.1
openpyxl==2.5.9
paramiko==2.4.2
pexpect==4.2.1
Pillow==5.3.0
protobuf==3.0.0
pyasn1==0.4.4
pycairo==1.16.2
pycparser==2.19
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.2.0
pygobject==3.26.1
PyJWT==1.6.4
pymacaroons==0.13.0
PyMySQL==0.9.2
PyNaCl==1.1.2
pyRFC3339==1.0
python-debian==0.1.32
pytz==2018.5
pyxdg==0.25
PyYAML==3.12
redis==2.10.6
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
require==0.1.1
scapy==2.4.0
SecretStorage==2.3.1
simplejson==3.13.2
six==1.11.0
tablib==0.12.1
tornado==5.1.1
ufw==0.35
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4
wadllib==1.3.2
xlrd==1.1.0
xlwt==1.3.0
xxhash==0.6.1
zope.interface==4.3.2
